requirements.txt
faiss-cpu
sentence-transformers
pypdf
numpy
openai(not used due to errors in free model)